+++ src/Migrate/Migrate.Autorest/custom/New-AzMigrateLocalServerReplication.ps1
@@ -86,7 +86,7 @@ function New-AzMigrateLocalServerReplication {
         [Parameter(Mandatory)]
         [Microsoft.Azure.PowerShell.Cmdlets.Migrate.Category('Path')]
         [System.String]
-        # Specifies the target Resource Group Id where the migrated VM resources will reside.
+        # Specifies the target resource group ARM ID where the migrated VM resources will reside.
         ${TargetResourceGroupId},
 
         [Parameter(Mandatory)]
@@ -98,7 +98,7 @@ function New-AzMigrateLocalServerReplication {
         [Parameter(ParameterSetName = 'ByIdDefaultUser', Mandatory)]
         [Microsoft.Azure.PowerShell.Cmdlets.Migrate.Category('Path')]
         [System.String]
-        # Specifies the Operating System disk for the source server to be migrated.
+        # Specifies the operating system disk for the source server to be migrated.
         ${OSDiskID},
 
         [Parameter(Mandatory)]
@@ -181,23 +181,39 @@ function New-AzMigrateLocalServerReplication {
             $isDynamicRamEnabled = [System.Convert]::ToBoolean($IsDynamicMemoryEnabled)
         }
         $HasTargetVMRam = $PSBoundParameters.ContainsKey('TargetVMRam')
+        $HasTargetVirtualSwitchId = $PSBoundParameters.ContainsKey('TargetVirtualSwitchId')
         $HasTargetTestVirtualSwitchId = $PSBoundParameters.ContainsKey('TargetTestVirtualSwitchId')
         $parameterSet = $PSCmdlet.ParameterSetName
-        
-        $MachineIdArray = $MachineId.Split("/")
-        if ($MachineIdArray.Length -lt 11) {
-            throw "Invalid machine ARM ID '$MachineId'"
+
+        # Validate ARM ID format from inputs
+        if (!(Test-AzureResourceIdFormat -Data $MachineId -Format $IdFormats.MachineArmIdTemplate)) {
+            throw "Invalid -MachineId '$MachineId'. A valid machine ARM ID should follow the format '$($IdFormats.MachineArmIdTemplate)'."
         }
+
+        if (!(Test-AzureResourceIdFormat -Data $TargetStoragePathId -Format $IdFormats.StoragePathArmIdTemplate)) {
+            throw "Invalid -TargetStoragePathId '$TargetStoragePathId'. A valid storage path ARM ID should follow the format '$($IdFormats.StoragePathArmIdTemplate)'."
+        }
+
+        if (!(Test-AzureResourceIdFormat -Data $TargetResourceGroupId -Format $IdFormats.ResourceGroupArmIdTemplate)) {
+            throw "Invalid -TargetResourceGroupId '$TargetResourceGroupId'. A valid resource group ARM ID should follow the format '$($IdFormats.ResourceGroupArmIdTemplate)'."
+        }
+
+        if ($HasTargetVirtualSwitchId -and !(Test-AzureResourceIdFormat -Data $TargetVirtualSwitchId -Format $IdFormats.LogicalNetworkArmIdTemplate)) {
+            throw "Invalid -TargetVirtualSwitchId '$TargetVirtualSwitchId'. A valid logical network ARM ID should follow the format '$($IdFormats.LogicalNetworkArmIdTemplate)'."
+        }
+
+        if ($HasTargetTestVirtualSwitchId -and !(Test-AzureResourceIdFormat -Data $TargetTestVirtualSwitchId -Format $IdFormats.LogicalNetworkArmIdTemplate)) {
+            throw "Invalid -TargetTestVirtualSwitchId '$TargetTestVirtualSwitchId'. A valid logical network ARM ID should follow the format '$($IdFormats.LogicalNetworkArmIdTemplate)'."
+        }
+
+        # Extract information from $MachineId
+        $MachineIdArray = $MachineId.Split("/")
         $SiteType = $MachineIdArray[7]
         $SiteName = $MachineIdArray[8]
         $ResourceGroupName = $MachineIdArray[4]
         $MachineName = $MachineIdArray[10]
 
         # Get the source site and the discovered machine
-        if (($SiteType -ne $SiteTypes.HyperVSites) -and ($SiteType -ne $SiteTypes.VMwareSites)) {
-            throw "Site type is not supported. Site type '$SiteType'. Check MachineId provided."
-        }
-        
         if ($SiteType -eq $SiteTypes.HyperVSites) {
             $instanceType = $AzLocalInstanceTypes.HyperVToAzLocal
 
@@ -318,11 +334,11 @@ function New-AzMigrateLocalServerReplication {
         }
         else
         {
-            throw "Unsupported site type '$SiteType'. Only Hyper-V and VMware sites are supported."
+            throw "Site type of '$SiteType' in -MachineId is not supported. Only '$($SiteTypes.HyperVSites)' and '$($SiteTypes.VMwareSites)' are supported."
         }
 
         if ([string]::IsNullOrEmpty($runAsAccountId)) {
-            throw "Unable to determine RunAsAccount for site '$SiteName' from machine '$MachineName'. Please verify your appliance setup."
+            throw "Unable to determine RunAsAccount for site '$SiteName' from machine '$MachineName'. Please verify your appliance setup and provided -MachineId."
         }
 
         # Validate the VM
@@ -560,7 +576,6 @@ function New-AzMigrateLocalServerReplication {
         $customProperties.TargetHciClusterId = $targetClusterId
         $customProperties.TargetResourceGroupId = $TargetResourceGroupId
         $customProperties.TargetVMName = $TargetVMName
-        $customProperties.TargetCpuCore = if ($HasTargetVMCPUCore) { $TargetVMCPUCore } else { $machine.NumberOfProcessorCore }
         $customProperties.IsDynamicRam = if ($HasIsDynamicMemoryEnabled) { $isDynamicRamEnabled } else {  $isSourceDynamicMemoryEnabled }
     
         # Determine target VM Hyper-V Generation
@@ -573,11 +588,37 @@ function New-AzMigrateLocalServerReplication {
             $customProperties.HyperVGeneration = if ($machine.Firmware -ieq "BIOS") { "1" } else { "2" }
         }
 
+        # Validate TargetVMCPUCore
+        if ($HasTargetVMCPUCore)
+        {
+            if ($TargetVMCPUCore -lt 1 -or $TargetVMCPUCore -gt 64)
+            {
+                throw "Specify -TargetVMCPUCore between 1 and 64."
+            }
+            $customProperties.TargetCpuCore = $TargetVMCPUCore
+        }
+        else
+        {
+            $customProperties.TargetCpuCore = $machine.NumberOfProcessorCore
+        }
+
         # Validate TargetVMRam
-        if ($HasTargetVMRam) {
-            # TargetVMRam needs to be greater than 0
-            if ($TargetVMRam -le 0) {
-                throw "Specify a target RAM that is greater than 0"    
+        if ($HasTargetVMRam)
+        {
+            if ($customProperties.HyperVGeneration -eq "1") {
+                # Between 512 MB and 1 TB
+                if ($TargetVMRam -lt 512 -or $TargetVMRam -gt 1048576)
+                {
+                    throw "Specify -TargetVMRAM between 512 and 1048576 MB (i.e., 1 TB) for Hyper-V Generation 1 VM."
+                }
+            }
+            else # Hyper-V Generation 2
+            {
+                # Between 32 MB and 12 TB
+                if ($TargetVMRam -lt 32 -or $TargetVMRam -gt 12582912)
+                {
+                    throw "Specify -TargetVMRAM between 32 and 12582912 MB (i.e., 12 TB) for Hyper-V Generation 2 VM."
+                }
             }
 
             $customProperties.TargetMemoryInMegaByte = $TargetVMRam 
@@ -604,21 +645,11 @@ function New-AzMigrateLocalServerReplication {
                 if ($null -eq $osDisk) {
                     throw "No Disk found with InstanceId $OSDiskID from discovered machine disks."
                 }
-
-                $diskName = Split-Path $osDisk.Path -leaf
-                if (IsReservedOrTrademarked($diskName)) {
-                    throw "The disk name $diskName or part of the name is a trademarked or reserved word."
-                }
             }
             elseif ($SiteType -eq $SiteTypes.VMwareSites) {  
                 $osDisk = $machine.Disk | Where-Object { $_.Uuid -eq $OSDiskID }
                 if ($null -eq $osDisk) {
                     throw "No Disk found with Uuid $OSDiskID from discovered machine disks."
-                }
-
-                $diskName = Split-Path $osDisk.Path -leaf
-                if (IsReservedOrTrademarked($diskName)) {
-                    throw "The disk name $diskName or part of the name is a trademarked or reserved word."
                 }
             }
 
@@ -663,14 +694,14 @@ function New-AzMigrateLocalServerReplication {
             # Validate DiskToInclude
             [PSCustomObject[]]$uniqueDisks = @()
             foreach ($disk in $DiskToInclude) {
-                # VHD is not supported as OS disk in Gen2 VMs
-                if ($customProperties.HyperVGeneration -eq "2" -and $disk.DiskFileFormat -eq "VHD" -and $disk.IsOSDisk) {
-                    throw "VHD disks cannot be used as the OS disk of Hyper-V Generation 2 VMs. Please replace disk with id '$($disk.DiskId)' in -DiskToInclude by re-running New-AzMigrateLocalDiskMappingObject with 'VHDX' as Format."
+                # Enforce VHDX for Gen2 VMs
+                if ($customProperties.HyperVGeneration -eq "2" -and $disk.DiskFileFormat -eq "VHD") {
+                    throw "Please specify 'VHDX' as Format for the disk with id '$($disk.DiskId)' in -DiskToInclude by re-running New-AzMigrateLocalDiskMappingObject."
                 }
 
-                # PhysicalSectorSize must be 512 for VHD format
-                if ($disk.DiskFileFormat -eq "VHD" -and $disk.DiskPhysicalSectorSize -ne 512) {
-                    throw "PhysicalSectorSize must be 512 for VHD format. Please replace disk with id '$($disk.DiskId)' in -DiskToInclude by re-running New-AzMigrateLocalDiskMappingObject with 512 as PhysicalSectorSize."
+                # PhysicalSectorSize must be 512 for VHD format if it is set
+                if ($disk.DiskFileFormat -eq "VHD" -and $null -ne $disk.DiskPhysicalSectorSize -and $disk.DiskPhysicalSectorSize -ne 512) {
+                    throw "Invalid Physical sector size of $($disk.DiskPhysicalSectorSize) is found for VHD format. Please replace disk with id '$($disk.DiskId)' in -DiskToInclude by re-running New-AzMigrateLocalDiskMappingObject with 512 as -PhysicalSectorSize."
                 }
 
                 if ($SiteType -eq $SiteTypes.HyperVSites) {
@@ -684,11 +715,6 @@ function New-AzMigrateLocalServerReplication {
                     if ($null -eq $discoveredDisk) {
                         throw "No Disk found with Uuid '$($disk.DiskId)' from discovered machine disks."
                     }
-                }
-
-                $diskName = Split-Path -Path $discoveredDisk.Path -Leaf
-                if (IsReservedOrTrademarked($diskName)) {
-                    throw "The disk name $diskName or part of the name is a trademarked or reserved word."
                 }
 
                 if ($uniqueDisks.Contains($disk.DiskId)) {